Move xrepo base into shared directory.
